+++ core-ng/src/main/java/core/framework/internal/web/HTTPIOHandler.java
@@ -36,6 +36,7 @@ public class HTTPIOHandler implements HttpHandler {
     @Override
     public void handleRequest(HttpServerExchange exchange) throws Exception {
         if (HEALTH_CHECK_PATH.equals(exchange.getRequestPath())) {      // not treat health-check as action
+            handler.addKeepAliveHeader(exchange);
             exchange.endExchange(); // end exchange will send 200 / content-length=0
             return;
         }